Rename CompleteParamDocblockFromToManyRector to CompleteParamDocblockFromSetterToCollectionRector as it works with setter
