+++ rules-tests/TypedCollections/Rector/Class_/CompleteParamDocblockFromSetterToCollectionRector/CompleteParamDocblockFromSetterToCollectionRectorTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Rector\Doctrine\Tests\TypedCollections\Rector\Class_\CompleteParamDocblockFromToManyRector;
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Class_\CompleteParamDocblockFromSetterToCollectionRector;
 
 use Iterator;
 use PHPUnit\Framework\Attributes\DataProvider;
 use Rector\Testing\PHPUnit\AbstractRectorTestCase;
 
-final class CompleteParamDocblockFromToManyRectorTest extends AbstractRectorTestCase
+final class CompleteParamDocblockFromSetterToCollectionRectorTest extends AbstractRectorTestCase
 {
     #[DataProvider('provideData')]
     public function test(string $filePath): void
