+++ rules-tests/TypedCollections/Rector/Class_/CompleteParamDocblockFromSetterToCollectionRector/Source/Training.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\Doctrine\Tests\TypedCollections\Rector\Class_\CompleteParamDocblockFromToManyRector\Source;
+namespace Rector\Doctrine\Tests\TypedCollections\Rector\Class_\CompleteParamDocblockFromSetterToCollectionRector\Source;
 
 use Doctrine\ORM\Mapping as ORM;
