--- rules-tests/TypedCollections/Rector/Class_/CompleteParamDocblockFromToManyRector/config/configured_rule.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\Doctrine\TypedCollections\Rector\Class_\CompleteParamDocblockFromToManyRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->rules([CompleteParamDocblockFromToManyRector::class]);
-};
